# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: StudyFlow AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /study-materials:
    get:
      operationId: listStudyMaterials
      tags: [studymaterials]
      summary: List all study materials
      responses:
        "200":
          description: List of study materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudyMaterial"
    post:
      operationId: createStudyMaterial
      tags: [studymaterials]
      summary: Upload and process a study material
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudyMaterialBody"
      responses:
        "201":
          description: Created study material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyMaterial"
  /study-materials/{id}:
    get:
      operationId: getStudyMaterial
      tags: [studymaterials]
      summary: Get a specific study material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Study material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyMaterial"
    delete:
      operationId: deleteStudyMaterial
      tags: [studymaterials]
      summary: Delete a study material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /study-materials/{id}/summarize:
    post:
      operationId: summarizeStudyMaterial
      tags: [studymaterials]
      summary: Generate AI summary and notes for a study material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: SSE stream of summary
          content:
            text/event-stream: {}
  /study-materials/{id}/quiz:
    post:
      operationId: generateQuiz
      tags: [studymaterials]
      summary: Generate quiz questions for a study material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateQuizBody"
      responses:
        "200":
          description: Generated quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /study-materials/{id}/exercises:
    post:
      operationId: generateExercises
      tags: [studymaterials]
      summary: Generate extra exercises for a study material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Generated exercises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exercise"
...
components:
  schemas:
...
    StudyMaterial:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        subject:
          type: string
        content:
          type: string
        summary:
          type: string
        fileType:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - subject
        - content
        - fileType
        - createdAt
        - updatedAt
    CreateStudyMaterialBody:
      type: object
      properties:
        title:
          type: string
        subject:
          type: string
        content:
          type: string
        fileType:
          type: string
      required:
        - title
        - subject
        - content
        - fileType
    GenerateQuizBody:
      type: object
      properties:
        difficulty:
          type: string
          enum: [makkelijk, gemiddeld, moeilijk]
        numQuestions:
          type: integer
      required:
        - difficulty
    Quiz:
      type: object
      properties:
        id:
          type: integer
        materialId:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - materialId
        - questions
        - createdAt
    QuizQuestion:
      type: object
      properties:
        question:
          type: string
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
        explanation:
          type: string
      required:
        - question
        - correctAnswer
    Exercise:
      type: object
      properties:
        id:
          type: integer
        materialId:
          type: integer
        question:
          type: string
        answer:
          type: string
        difficulty:
          type: string
        tip:
          type: string
      required:
        - id
        - materialId
        - question
        - answer
        - difficulty
...