# lib/api-zod/src/generated/types/createStudyMaterialBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudyFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateStudyMaterialBody {
  title: string;
  subject: string;
  content: string;
  fileType: string;
}