# lib/api-zod/src/generated/types/quiz.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudyFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuizQuestion } from "./quizQuestion";

export interface Quiz {
  id: number;
  materialId: number;
  questions: QuizQuestion[];
  createdAt: Date;
}